refactor: canvases/ を feat-recruit/common/canvases/ に移動
